feat: merge cchecks, SD-1248
