+++ .github/actions/repo-checks/templates/pr-comment-secrets-template.j2
