+++ .github/actions/repo-checks/templates/slack-message-template.j2
