--- .github/actions/vulnerability-checks/action.yaml
@@ -1,90 +1,0 @@
-name: Vulnerability-checks
-description: This check finds the vulnerabilities in the code
-inputs:
-  enable-vulnerability-checks:
-    description: |
-      Enable vulnerabilities check
-      https://aquasecurity.github.io/trivy/
-    required: true
-    default: "false"
-  trivy-version:
-    description: |
-      Trivy version to install
-    required: false
-    default: "v0.61.0"
-  enable-pr-comment:
-    description: |
-      Enable pr comment
-    required: true
-    default: "true"
-  # GitHub token used to authenticate API requests.
-  # Use the default GitHub Actions token: ${{ secrets.GITHUB_TOKEN }}.
-  github-token:
-    description: Github token secret
-    required: true
-runs:
-  using: composite
-  steps:
-    - name: Exit if checks are disabled
-      if: inputs.enable-vulnerability-checks == 'false'
-      shell: bash
-      run: |
-        exit 0
-    - name: Setup environment
-      uses: saritasa-nest/saritasa-github-actions/.github/actions/install-checks-dependencies@feature/checks-test
-      with: 
-        trivy-version: ${{inputs.trivy-version}}
-    - name: Run trivy vulnerability scanner in fs mode
-      uses: aquasecurity/trivy-action@0.30.0
-      id: trivy-vulnerability-checks
-      with:
-        scan-type: fs
-        scan-ref: .
-        format: sarif
-        exit-code: 1
-        output: vulnerabilities-results.sarif
-        skip-setup-trivy: true
-        scanners: vuln
-        severity: HIGH,CRITICAL
-      env:
-        TRIVY_DB_REPOSITORY: "public.ecr.aws/aquasecurity/trivy-db:2"
-    - name: Add trivy failure vulnerabilities check results to data file
-      shell: bash
-      if: always() && steps.trivy-vulnerability-checks.conclusion == 'failure'
-      run: |
-        python3 ${{ github.action_path }}/scripts/sarif-to-json.py vulnerabilities-results.sarif vulnerabilities-output.json vulnerabilities
-        # Add the contents of the vulnerabilities-output.json file to data.json, from which 
-        # notifications about the results of vulnerability checks are generated
-        jq -s '.[0] * .[1]' vulnerabilities-output.json data.json | sponge data.json
-        echo reactions="confused" > $GITHUB_ENV
-    - name: Create pr comment with checks results
-      shell: bash
-      if: always()
-      run: |
-        python3 ${{ github.action_path }}/scripts/generate-message.py \
-          ${{ github.action_path }}/templates/pr-comment-template.j2 \
-          data.json \
-          pr-comment.md
-    - name: Add comment to github actions summary
-      if: always()
-      shell: bash
-      run: |
-        cat pr-comment.md > $GITHUB_STEP_SUMMARY
-    - name: Find pr comment with check results
-      uses: peter-evans/find-comment@v3
-      if: always() && github.event_name == 'pull_request' && inputs.enable-pr-comment == 'true'
-      id: find-comment
-      with:
-        issue-number: ${{ github.event.number }}
-        comment-author: "github-actions[bot]"
-        body-includes: Summary of the vulnerabilities check
-    - name: Upload checks results as a pr comment
-      if: always() && github.event_name == 'pull_request' && inputs.enable-pr-comment == 'true'
-      uses: peter-evans/create-or-update-comment@v4
-      with:
-        issue-number: ${{ github.event.number }}
-        comment-id: ${{ steps.find-comment.outputs.comment-id }}
-        body-path: pr-comment.md
-        edit-mode: replace
-        reactions: ${{ env.reactions || 'laugh' }}
-        reactions-edit-mode: replace
